+++ packages/antd-ext/components/CascaderExt/index.ts
@@ -1,0 +1,5 @@
+import BaseCascaderExt from './CascaderExt';
+
+export {
+  BaseCascaderExt,
+}